Move RktGc to DomainMgr and run it in a go routine instead of main thread to avoid watchdogs.
